+++ contracts/Core/storage/BlockStorage.sol
@@ -3,17 +3,17 @@ pragma solidity ^0.8.0;
 import "../../lib/Structs.sol";
 
 contract BlockStorage {
-    //epoch -> address -> dispute -> assetid
+    //epoch -> address -> dispute
     mapping(uint32 => mapping(address => Structs.Dispute)) public disputes;
     //epoch -> blockId -> block
-    mapping(uint32 => mapping(uint8 => Structs.Block)) public proposedBlocks;
+    mapping(uint32 => mapping(uint16 => Structs.Block)) public proposedBlocks;
     //epoch->blockId
-    mapping(uint32 => uint8[]) public sortedProposedBlockIds;
+    mapping(uint32 => uint16[]) public sortedProposedBlockIds;
     //stakerId->epoch
     mapping(uint32 => uint32) public epochLastProposed;
 
     // slither-disable-next-line constable-states
-    uint8 public numProposedBlocks;
+    uint16 public numProposedBlocks;
     // slither-disable-next-line constable-states
     int8 public blockIndexToBeConfirmed; // Index in sortedProposedBlockIds
     // epoch -> blocks
